+++ src/controllers/RedirectsController.php
@@ -5,11 +5,16 @@ namespace studioespresso\seofields\controllers;
 use Craft;
 use craft\helpers\App;
 use craft\helpers\Cp;
+use craft\helpers\DateTimeHelper;
 use craft\helpers\UrlHelper;
 use craft\models\Site;
+use craft\services\Path;
 use craft\web\Controller;
-use League\Csv\Reader;
+use OpenSpout\Common\Entity\Row;
+use OpenSpout\Reader\CSV\Reader as CsvReader;
+use OpenSpout\Writer\XLSX\Writer;
 use studioespresso\seofields\models\RedirectModel;
+use studioespresso\seofields\records\RedirectRecord;
 use studioespresso\seofields\SeoFields;
 use yii\web\UploadedFile;
 
@@ -104,26 +109,51 @@ class RedirectsController extends Controller
     {
         $this->requirePostRequest();
 
-        // If your CSV document was created or is read on a Macintosh computer,
-        // add the following lines before using the library to help PHP detect line ending in Mac OS X
-        if (!ini_get('auto_detect_line_endings')) {
-            ini_set('auto_detect_line_endings', '1');
-        }
-
         $file = UploadedFile::getInstanceByName('file');
 
         if ($file !== null) {
             $filename = self::IMPORT_FILE;
             $filePath = Craft::$app->getPath()->getTempPath() . DIRECTORY_SEPARATOR . $filename;
             $file->saveAs($filePath, false);
-            $csv = Reader::createFromPath($file->tempName);
-            $headers = $csv->fetchOne(0);
-            Craft::info(print_r($headers, true), __METHOD__);
-            $variables['headers'] = $headers;
-            $variables['filename'] = $filePath;
         }
 
         $this->redirect(UrlHelper::cpUrl('seo-fields/redirects/import'));
+    }
+
+    public function actionExport()
+    {
+        $site = $this->request->getQueryParam('site', null);
+        /** @var Path $pathService */
+        $pathService = Craft::$app->getPath();
+        $now = DateTimeHelper::now();
+        $path = $pathService->getTempPath() . "/redirect-{$now->format('Y-m-d h:i:s')}.xlsx";
+
+        $writer = new Writer();
+        $writer->openToFile($path);
+
+        $headerRow = Row::fromValues(["Old url", "Redirected to", "Type", "Site Name", "Last hit on", "Total hits"]);
+
+        $writer->addRow($headerRow);
+        if ($site) {
+            $site = Craft::$app->getSites()->getSiteByHandle($site);
+            $redirects = RedirectRecord::findAll(['siteId' => $site->id]);
+        } else {
+            $redirects = RedirectRecord::find()->all();
+        }
+        /** @var RedirectRecord[] $redirects */
+        foreach ($redirects as $redirect) {
+            $row = Row::fromValues([
+                $redirect->pattern,
+                $redirect->redirect,
+                $redirect->method,
+                $redirect->siteId ? Craft::$app->getSites()->getSiteById($redirect->siteId)->name : 'All Sites',
+                $redirect->dateLastHit ? DateTimeHelper::toDateTime($redirect->dateLastHit)->format('Y-m-d h:i:s') : '',
+                $redirect->counter,
+            ]);
+            $writer->addRow($row);
+        }
+        $writer->close();
+        return Craft::$app->getResponse()->sendFile($path);
     }
 
     public function actionImport()
@@ -133,8 +163,7 @@ class RedirectsController extends Controller
         if (!file_exists($filePath)) {
             return $this->redirect(UrlHelper::cpUrl('seo-fields/redirects'));
         }
-        $csv = Reader::createFromPath($filePath);
-        $headers = $csv->fetchOne(0);
+        $headers = $this->getHeaders($filePath);
         $variables['headers'] = $headers;
         $variables['filename'] = $filePath;
         $variables['sites'] = $this->getSitesMenu();
@@ -159,14 +188,9 @@ class RedirectsController extends Controller
 
         $filename = self::IMPORT_FILE;
         $filePath = Craft::$app->getPath()->getTempPath() . DIRECTORY_SEPARATOR . $filename;
-        $reader = Reader::createFromPath($filePath);
-        //$headers = $reader->fetchOne(0);
-        $reader->setHeaderOffset(1);
 
-        $headers = $this->getHeaders($reader);
-        $rows = $this->getRows($reader);
-        $variables['headers'] = $headers;
-        $variables['filename'] = $filePath;
+        $rows = $this->getRows($filePath);
+        $headers = $this->getHeaders($filePath);
 
         $results = SeoFields::getInstance()->redirectService->import($rows, $settings);
         return $this->renderTemplate('seo-fields/_redirect/_import_results', $results);
@@ -209,32 +233,40 @@ class RedirectsController extends Controller
         return $sites;
     }
 
-    private function getHeaders($reader)
+    private function getHeaders(string $filePath): array
     {
-        // Support for league/csv v8 with a header
-        try {
-            return $reader->fetchOne(0);
-        } catch (\Throwable $e) {
+        $reader = new CsvReader();
+        $reader->open($filePath);
+
+        foreach ($reader->getSheetIterator() as $sheet) {
+            foreach ($sheet->getRowIterator() as $row) {
+                $reader->close();
+                return $row->toArray();
+            }
         }
 
-        try {
-            $reader->setHeaderOffset(0);
-            return $reader->getHeader();
-        } catch (\Throwable $e) {
-        }
+        $reader->close();
+        return [];
     }
 
-    private function getRows(Reader $reader)
+    private function getRows(string $filePath): array
     {
-        try {
-            /** @phpstan-ignore-next-line */
-            return $reader->fetchAll();
-        } catch (\Throwable $e) {
+        $reader = new CsvReader();
+        $reader->open($filePath);
+
+        $rows = [];
+        $isFirstRow = true;
+        foreach ($reader->getSheetIterator() as $sheet) {
+            foreach ($sheet->getRowIterator() as $row) {
+                if ($isFirstRow) {
+                    $isFirstRow = false;
+                    continue;
+                }
+                $rows[] = $row->toArray();
+            }
         }
 
-        try {
-            return $reader->getIterator();
-        } catch (\Throwable $e) {
-        }
+        $reader->close();
+        return $rows;
     }
 }